Fix naming conventions and remove unnecessary code from notifications.ts
